made to run more easily
delete all old stuff. moved function to the top of the script. now if
you check the “source on save” box at the top of the script and push
save, everything will run as if “sourced” like matlab. common errors
from this approach will be due to incorrect setwd() so make sure its
right.